fix: add each modifier to apply
