+++ demo/index.html
@@ -16,7 +16,7 @@
   <div :text.trim="state.test" myHidden="true"></div>
 
   <div :text="state.value"></div>
-
+  
   <div :text.unit.$="state.value"></div>
 
   <div :text.fixed.2="state.test2"></div>
@@ -27,15 +27,15 @@
 
   <div :onclick="state.value++">add</div>
 
-<!--  <div :onclick="state.arr.splice(1, 0, { color: 'black' })">slice</div>-->
-<!--  <div :onclick="state.arr.push({ color: 'black' })">push</div>-->
-<!--  <div :onclick="state.arr.shift()">shift</div>-->
-<!--  <div :onclick="state.arr.unshift({ color: 'black' })">unshift</div>-->
-<!--  <div :onclick="state.arr[2].color = 'pink'">change</div>-->
-<!--  <hr>-->
-<!--  <div :each="state.arr">-->
-<!--      <div :text="item.color" class="color" template></div>-->
-<!--  </div>-->
+<div :onclick="state.arr.splice(1, 0, { color: 'black' })">slice</div>
+<div :onclick="state.arr.push({ color: 'black' })">push</div>
+<div :onclick="state.arr.shift()">shift</div>
+<div :onclick="state.arr.unshift({ color: 'black' })">unshift</div>
+<div :onclick="state.arr[2].color = 'pink'">change</div>
+<hr>
+<div :each="state.arr">
+<div :text="item.color" class="color" template></div>
+</div>
   
   <script type="module">
