--- src/directives/methods/_html.js
@@ -1,12 +1,0 @@
-import { applyModifier } from "../modifiers/applyModifier"
-
-/**
- *
- * @param {HTMLElement} element
- * @param {any} value
- * @param {string} modifier
- * @param {string | number} args
- */
-export function _html(element, value, modifier, args) {
-  element.innerHTML = applyModifier(value, modifier, args) ?? ""
-}
