--- src/directives/methods/_show.js
@@ -1,8 +1,0 @@
-/**
- *
- * @param {HTMLElement} element
- * @param {boolean} value
- */
-export function _show(element, value) {
-  element.style.display = value ? "" : "none"
-}
