--- src/directives/methods/_text.js
@@ -1,11 +1,0 @@
-import { applyModifier } from "../modifiers/applyModifier";
-/**
- *
- * @param {HTMLElement} element
- * @param {any} value
- * @param {string} modifier
- * @param {string | number} args
- */
-export function _text(element, value, modifier, args) {
-  element.textContent = applyModifier(value, modifier, args) ?? ""
-}
