+++ src/directives/methods/index.js
@@ -1,17 +1,17 @@
-import { _class } from "./_class.js"
-import { _text } from "./_text.js"
-import { _each } from "./_each.js"
-import { _attr } from "./_attr.js"
-import { _html } from "./_html.js"
-import { _show } from "./_show.js"
-import { _style } from "./_style.js"
+import { classes } from './classes.js'
+import { text } from './text.js'
+import { each } from './each.js'
+import { attr } from './attr.js'
+import { html } from './html.js'
+import { show } from './show.js'
+import { style } from './style.js'
 
 export default {
-  _class,
-  _text,
-  _each,
-  _html,
-  _show,
-  _style,
-  
+	classes,
+	text,
+	each,
+	html,
+	show,
+	style,
+	attr
 }
