+++ src/directives/methods/show.js
@@ -1,0 +1,8 @@
+/**
+ *
+ * @param {HTMLElement} element
+ * @param {boolean} value
+ */
+export function show(element, value) {
+	element.style.display = value ? '' : 'none'
+}
